requirements.txt
torch>=2.0.0
transformers>=4.35.0
pillow>=10.0.0
tqdm>=4.66.0
numpy>=1.24.0
sqlite-vec>=0.0.1
sentencepiece>=0.1.99
umap-learn>=0.5.5
plotly>=5.18.0
pandas>=2.0.0